gateway: split read-side callbacks into NativeObservabilityMixin
native_bridge.py crossed the repository line budget; the content-free
read-side callbacks (granted-model discovery, usage report, metrics
snapshot, readiness, per-worker resource release) move verbatim into
exp/runtime/gateway/native_observability.py as a stateless mixin the
control plane inherits, so the bridge module keeps only the
request-serving authority path. No behavior change.
