+++ exp/runtime/gateway/native_observability_test.py
@@ -1,0 +1,1 @@
+"""Covered by exp/runtime/gateway/native_bridge_test.py through the bridge boundary."""
